requirements.txt
Django
gevent
gevent-socketio-hartwork
